mqtt/requirements.txt
appdirs==1.4.4
bidict==0.21.2
black==21.5b1
certifi==2020.12.5
click==8.0.1
dnspython==1.16.0
dotobject==1.2.2
eventlet==0.31.0
Flask==2.0.1
Flask-Cors==3.0.9
Flask-SocketIO==5.0.1
greenlet==1.1.0
hiredis==2.0.0
influxdb-client==1.17.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
paho-mqtt==1.5.1
pathspec==0.8.1
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
redis==3.5.3
redisworks==0.3.0
regex==2021.4.4
Rx==3.2.0
six==1.16.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.0.1
